requirements.txt
alembic==1.15.2
babel==2.17.0
bcrypt==4.3.0
chardet==5.2.0
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
et_xmlfile==2.0.0
fonttools==4.57.0
greenlet==3.2.1
kiwisolver==1.4.8
logging==0.4.9.6
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.0
six==1.17.0
SQLAlchemy==2.0.40
tk==0.1.0
tkcalendar==1.6.1
typing_extensions==4.13.2
tzdata==2025.2
